backend/requirements.txt
elevenlabs>=1.0.0
twilio>=9.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0.0
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0

